+++ CHANGELOG.md
@@ -12,3 +12,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Enable match that matches non-null references (#11)
 - Enable match that matches equality using an equality comparer (#7)
 - Enable match that matches inequality using an equality comparer (#9)
+- Enable match that matches same instances (#13)